playercurrent to return nbmeeples for the current player too

+++ Appli_serveur_test/Appli_serveur_test/Player.cs
@@ -16,7 +16,7 @@ namespace system
         public bool _is_ready { get; set; }
         public Socket? _socket_of_player { get; set; }
 
-        public int _nbMeeples { get; set; }
+        public ulong _nbMeeples { get; set; }
 
         public Semaphore _s_player;
 
@@ -38,7 +38,7 @@ namespace system
             _s_player = new Semaphore(1, 1);
         }
 
-        public Player(ulong id_player, Socket? playerSocket, int nbMeeples)
+        public Player(ulong id_player, Socket? playerSocket, ulong nbMeeples)
         {
             _id_player = id_player;
             _score = 0;
